columnas "cmd" y "unidad" nuevas en plantas front y back

+++ backend/Coberturas/Coberturas/Models/Trades/Banks.cs
@@ -5,6 +5,6 @@ namespace Coberturas.Models.Trades
     public int id_bank { get; set; }
     public string bank { get; set; }
     public int CSA { get; set; }
-    public float threshold { get; set; }
+    public double threshold { get; set; }
   }
 }
